[feedback] Refactor Agent Feedback History Collection & Formatting
## What
Shifted the agent event history formatting logic from the graph editing agent action module to the global feedback controller, modifying the feedback API to accept raw event histories instead of pre-formatted strings, and added support for formatting multiple agent sessions.

## Why
Enables other types of feedback to easily include agent event histories. Centralizing this formatting inside `FeedbackController` keeps action coordinators clean and allows the feedback system to handle multiple agent sessions (e.g. graph editing and other agent sessions in the same graph) uniformly.

## Changes

### 1. Centralized Formatting in Feedback Controller
- **Location**: [feedback-controller.ts](file:///Users/dglazkov/Documents/code/breadboard/packages/visual-editor/src/sca/controller/subcontrollers/global/feedback-controller.ts)
- Imported `AgentEvent` and associated event payload helpers.
- Added private module-level helpers `formatLLMContent` and `formatAgentEventHistory`.
- Updated `open()` to accept an optional `agentEvents` array of histories: `agentEvents?: Array<ReadonlyArray<AgentEvent>>`.
- Formats and joins multiple histories with headers (e.g., `=== Agent Session #1 ===`) if more than one session is present, storing the resulting string under the `conversation` field in `productData`.

### 2. Action Simplification
- **Location**: [graph-editing-agent-actions.ts](file:///Users/dglazkov/Documents/code/breadboard/packages/visual-editor/src/sca/actions/agent/graph-editing-agent-actions.ts)
- Removed local `formatLLMContent` and `formatAgentEventHistory` functions and cleaned up unused imports.
- Updated `setOpieReaction` to pass the raw history array `[events]` via the `agentEvents` option of `feedback.open`.

### 3. Unit Test Coverage
- **Location**: [graph-editing-agent-actions.test.ts](file:///Users/dglazkov/Documents/code/breadboard/tests/sca/actions/agent/graph-editing-agent-actions.test.ts)
  - Updated the `setOpieReaction` test to assert that `feedback.open` is called with the raw `agentEvents` array instead of the pre-formatted `conversation` string.
- **Location**: [feedback-controller.test.ts](file:///Users/dglazkov/Documents/code/breadboard/tests/sca/controller/subcontrollers/global/feedback-controller.test.ts)
  - Updated the `startFeedback` mock to capture `providedProductData`.
  - Added new test cases to verify correct formatting and separation of single and multiple agent session histories.

## Testing
Verified that all tests compile and pass successfully:
- `npm run test:file -- './dist/tsc/tests/sca/controller/subcontrollers/global/feedback-controller.test.js'` (14/14 passed)
- `npm run test:file -- './dist/tsc/tests/sca/actions/agent/graph-editing-agent-actions.test.js'` (17/17 passed)
